app/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pandas
openpyxl
mail-parser